kuji-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
redis[asyncio]==5.2.1
python-jose[cryptography]==3.3.0
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1
